+++ src/main/java/com/lesson/repository/tables/EduCourseRecord.java
@@ -125,7 +125,7 @@ public class EduCourseRecord extends TableImpl<EduCourseRecordRecord> {
     /**
      * The column <code>lesson.edu_course_record.deleted</code>. 是否删除：0-未删除，1-已删除
      */
-    public final TableField<EduCourseRecordRecord, Byte> DELETED = createField(DSL.name("deleted"), SQLDataType.TINYINT.nullable(false).defaultValue(DSL.inline("0", SQLDataType.TINYINT)), this, "是否删除：0-未删除，1-已删除");
+    public final TableField<EduCourseRecordRecord, Integer> DELETED = createField(DSL.name("deleted"), SQLDataType.INTEGER.nullable(false).defaultValue(DSL.inline("0", SQLDataType.INTEGER)), this, "是否删除：0-未删除，1-已删除");
 
     private EduCourseRecord(Name alias, Table<EduCourseRecordRecord> aliased) {
         this(alias, aliased, null);
@@ -216,7 +216,7 @@ public class EduCourseRecord extends TableImpl<EduCourseRecordRecord> {
     // -------------------------------------------------------------------------
 
     @Override
-    public Row15<Long, String, String, String, LocalDateTime, LocalDateTime, BigDecimal, String, Long, String, Long, String, LocalDateTime, LocalDateTime, Byte> fieldsRow() {
+    public Row15<Long, String, String, String, LocalDateTime, LocalDateTime, BigDecimal, String, Long, String, Long, String, LocalDateTime, LocalDateTime, Integer> fieldsRow() {
         return (Row15) super.fieldsRow();
     }
 }
