+++ src/main/java/com/lesson/repository/tables/EduStudentClassTransfer.java
@@ -119,7 +119,7 @@ public class EduStudentClassTransfer extends TableImpl<EduStudentClassTransferRe
     /**
      * The column <code>lesson.edu_student_class_transfer.deleted</code>. 是否删除：0-未删除，1-已删除
      */
-    public final TableField<EduStudentClassTransferRecord, Byte> DELETED = createField(DSL.name("deleted"), SQLDataType.TINYINT.nullable(false).defaultValue(DSL.inline("0", SQLDataType.TINYINT)), this, "是否删除：0-未删除，1-已删除");
+    public final TableField<EduStudentClassTransferRecord, Integer> DELETED = createField(DSL.name("deleted"), SQLDataType.INTEGER.nullable(false).defaultValue(DSL.inline("0", SQLDataType.INTEGER)), this, "是否删除：0-未删除，1-已删除");
 
     private EduStudentClassTransfer(Name alias, Table<EduStudentClassTransferRecord> aliased) {
         this(alias, aliased, null);
@@ -210,7 +210,7 @@ public class EduStudentClassTransfer extends TableImpl<EduStudentClassTransferRe
     // -------------------------------------------------------------------------
 
     @Override
-    public Row14<Long, String, String, String, String, String, String, Long, String, Long, String, LocalDateTime, LocalDateTime, Byte> fieldsRow() {
+    public Row14<Long, String, String, String, String, String, String, Long, String, Long, String, LocalDateTime, LocalDateTime, Integer> fieldsRow() {
         return (Row14) super.fieldsRow();
     }
 }
