+++ src/main/java/com/lesson/repository/tables/EduStudentCourseTransfer.java
@@ -136,7 +136,7 @@ public class EduStudentCourseTransfer extends TableImpl<EduStudentCourseTransfer
     /**
      * The column <code>lesson.edu_student_course_transfer.deleted</code>. 是否删除：0-未删除，1-已删除
      */
-    public final TableField<EduStudentCourseTransferRecord, Byte> DELETED = createField(DSL.name("deleted"), SQLDataType.TINYINT.nullable(false).defaultValue(DSL.inline("0", SQLDataType.TINYINT)), this, "是否删除：0-未删除，1-已删除");
+    public final TableField<EduStudentCourseTransferRecord, Integer> DELETED = createField(DSL.name("deleted"), SQLDataType.INTEGER.nullable(false).defaultValue(DSL.inline("0", SQLDataType.INTEGER)), this, "是否删除：0-未删除，1-已删除");
 
     private EduStudentCourseTransfer(Name alias, Table<EduStudentCourseTransferRecord> aliased) {
         this(alias, aliased, null);
@@ -227,7 +227,7 @@ public class EduStudentCourseTransfer extends TableImpl<EduStudentCourseTransfer
     // -------------------------------------------------------------------------
 
     @Override
-    public Row17<Long, String, String, String, String, String, BigDecimal, BigDecimal, LocalDate, String, Long, String, Long, String, LocalDateTime, LocalDateTime, Byte> fieldsRow() {
+    public Row17<Long, String, String, String, String, String, BigDecimal, BigDecimal, LocalDate, String, Long, String, Long, String, LocalDateTime, LocalDateTime, Integer> fieldsRow() {
         return (Row17) super.fieldsRow();
     }
 }
