+++ src/main/java/com/lesson/repository/tables/EduStudentRefund.java
@@ -135,7 +135,7 @@ public class EduStudentRefund extends TableImpl<EduStudentRefundRecord> {
     /**
      * The column <code>lesson.edu_student_refund.deleted</code>. 是否删除：0-未删除，1-已删除
      */
-    public final TableField<EduStudentRefundRecord, Byte> DELETED = createField(DSL.name("deleted"), SQLDataType.TINYINT.nullable(false).defaultValue(DSL.inline("0", SQLDataType.TINYINT)), this, "是否删除：0-未删除，1-已删除");
+    public final TableField<EduStudentRefundRecord, Integer> DELETED = createField(DSL.name("deleted"), SQLDataType.INTEGER.nullable(false).defaultValue(DSL.inline("0", SQLDataType.INTEGER)), this, "是否删除：0-未删除，1-已删除");
 
     private EduStudentRefund(Name alias, Table<EduStudentRefundRecord> aliased) {
         this(alias, aliased, null);
@@ -226,7 +226,7 @@ public class EduStudentRefund extends TableImpl<EduStudentRefundRecord> {
     // -------------------------------------------------------------------------
 
     @Override
-    public Row17<Long, String, String, String, BigDecimal, BigDecimal, BigDecimal, BigDecimal, BigDecimal, String, Long, String, Long, String, LocalDateTime, LocalDateTime, Byte> fieldsRow() {
+    public Row17<Long, String, String, String, BigDecimal, BigDecimal, BigDecimal, BigDecimal, BigDecimal, String, Long, String, Long, String, LocalDateTime, LocalDateTime, Integer> fieldsRow() {
         return (Row17) super.fieldsRow();
     }
 }
